backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
langchain
langchain-google-genai
python-dotenv
